+++ arch/xtensa/src/esp32/esp32_wifi_utils.c
@@ -33,6 +33,7 @@
 
 #include "esp32_wifi_adapter.h"
 #include "esp32_wifi_utils.h"
+#include "esp32_wireless.h"
 #include "espidf_wifi.h"
 
 /****************************************************************************
