+++ arch/xtensa/src/esp32/esp32_wlan.h
@@ -40,7 +40,7 @@ extern "C"
 #define EXTERN extern
 #endif
 
-#ifdef CONFIG_ESP32_WIRELESS
+#ifdef CONFIG_ESP32_WIFI
 
 /****************************************************************************
  * Public Function Prototypes
@@ -82,7 +82,7 @@ int esp32_wlan_sta_initialize(void);
 int esp32_wlan_softap_initialize(void);
 #endif
 
-#endif /* CONFIG_ESP32_WIRELESS */
+#endif /* CONFIG_ESP32_WIFI */
 #ifdef __cplusplus
 }
 #endif
