+++ boards/xtensa/esp32/esp32-devkitc/configs/coex/defconfig
@@ -69,6 +69,7 @@ CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_SPIFLASH_SMARTFS=y
 CONFIG_ESP32_STORAGE_MTD_SIZE=0x80000
 CONFIG_ESP32_UART0=y
+CONFIG_ESP32_WIFI=y
 CONFIG_ESP32_WIFI_BT_COEXIST=y
 CONFIG_ESP32_WIFI_SAVE_PARAM=y
 CONFIG_EXAMPLE_POWER_SAVE_MIN_MODEM=y
