+++ boards/xtensa/esp32/esp32-devkitc/configs/sta_softap/defconfig
@@ -34,9 +34,9 @@ CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_SPIFLASH_SPIFFS=y
 CONFIG_ESP32_STORAGE_MTD_SIZE=0x80000
 CONFIG_ESP32_UART0=y
+CONFIG_ESP32_WIFI=y
 CONFIG_ESP32_WIFI_SAVE_PARAM=y
 CONFIG_ESP32_WIFI_STATION_SOFTAP_COEXISTENCE=y
-CONFIG_ESP32_WIRELESS=y
 CONFIG_EXAMPLES_DHCPD=y
 CONFIG_EXPERIMENTAL=y
 CONFIG_FS_PROCFS=y
