+++ boards/xtensa/esp32/esp32-devkitc/configs/wamr_wasi_debug/defconfig
@@ -49,7 +49,7 @@ CONFIG_ESP32_IRAM_HEAP=y
 CONFIG_ESP32_RTC_HEAP=y
 CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_UART0=y
-CONFIG_ESP32_WIRELESS=y
+CONFIG_ESP32_WIFI=y
 CONFIG_EXPERIMENTAL=y
 CONFIG_FS_LITTLEFS=y
 CONFIG_FS_PROCFS=y
