+++ boards/xtensa/esp32/esp32-devkitc/configs/wapi/defconfig
@@ -23,18 +23,9 @@ CONFIG_ARCH_XTENSA=y
 CONFIG_BOARD_LOOPSPERMSEC=16717
 CONFIG_BUILTIN=y
 CONFIG_DEBUG_ASSERTIONS=y
-CONFIG_DEBUG_ERROR=y
 CONFIG_DEBUG_FEATURES=y
 CONFIG_DEBUG_FULLOPT=y
-CONFIG_DEBUG_INFO=y
-CONFIG_DEBUG_MM=y
-CONFIG_DEBUG_MM_ERROR=y
-CONFIG_DEBUG_MM_WARN=y
-CONFIG_DEBUG_NET=y
-CONFIG_DEBUG_NET_ERROR=y
-CONFIG_DEBUG_NET_WARN=y
 CONFIG_DEBUG_SYMBOLS=y
-CONFIG_DEBUG_WARN=y
 CONFIG_DEFAULT_TASK_STACKSIZE=4096
 CONFIG_DEV_URANDOM=y
 CONFIG_DRIVERS_IEEE80211=y
@@ -43,8 +34,8 @@ CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_SPIFLASH_SPIFFS=y
 CONFIG_ESP32_STORAGE_MTD_SIZE=0x80000
 CONFIG_ESP32_UART0=y
+CONFIG_ESP32_WIFI=y
 CONFIG_ESP32_WIFI_SAVE_PARAM=y
-CONFIG_ESP32_WIRELESS=y
 CONFIG_FS_PROCFS=y
 CONFIG_IDLETHREAD_STACKSIZE=3072
 CONFIG_INIT_ENTRYPOINT="nsh_main"
