+++ boards/xtensa/esp32/esp32-devkitc/configs/wapi_smp/defconfig
@@ -34,8 +34,8 @@ CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_SPIFLASH_SPIFFS=y
 CONFIG_ESP32_STORAGE_MTD_SIZE=0x80000
 CONFIG_ESP32_UART0=y
+CONFIG_ESP32_WIFI=y
 CONFIG_ESP32_WIFI_SAVE_PARAM=y
-CONFIG_ESP32_WIRELESS=y
 CONFIG_EXAMPLES_HELLO=y
 CONFIG_EXAMPLES_WEBSERVER=y
 CONFIG_FS_PROCFS=y
