+++ boards/xtensa/esp32/esp32-wrover-kit/configs/mcuboot_update_agent/defconfig
@@ -33,7 +33,7 @@ CONFIG_DRIVERS_WIRELESS=y
 CONFIG_ESP32_APP_FORMAT_MCUBOOT=y
 CONFIG_ESP32_SPIFLASH=y
 CONFIG_ESP32_UART0=y
-CONFIG_ESP32_WIRELESS=y
+CONFIG_ESP32_WIFI=y
 CONFIG_EXAMPLES_MCUBOOT_UPDATE_AGENT=y
 CONFIG_EXAMPLES_MCUBOOT_UPDATE_AGENT_DL_BUFFER_SIZE=4096
 CONFIG_EXPERIMENTAL=y
